requirements.txt
python-telegram-bot==11.1.0
environs==9.5.0
redis==4.3.3
vk-api==11.9.8